Core,Web: Flags are now mainteined by the session
Before this commit flags were managed by each Exploit,
after this commit flags will be managed by Session,
this is an important structural change.

A FlagStore manages all the flags, MemoryFlagStore is similar to storing
the flags before this commit, all the flags are stored in the json
config.

The Flag now contains more information like from where the flag was
received, the current status and it's Service and Exploit